feat: frame subagent synthesis as background work when user has moved on
When a subagent result arrives after the user has already sent a new
message, the synthesis loop now detects the active user-message loop
and injects a system prompt telling the LLM to frame its response as
a follow-up from earlier background work rather than as a response to
the current topic. This prevents confusing interleaving of stale
subagent results with the user's current conversation.

Promotes SessionBackgroundTaskTracker from RockBot.Agent into the
framework (RockBot.Host) behind a new ISessionTracker interface so
any agent built on the framework gets this behavior automatically.
